moved assets according to convension
