+++ src/main.ts
@@ -4,9 +4,9 @@ import { DefaultApolloClient } from '@vue/apollo-composable'
 import { createPinia } from 'pinia'
 import ToastService from 'primevue/toastservice'
 import PrimeVue from 'primevue/config'
-import { iaPreset } from './assets/styles/themePreset'
+import { iaPreset } from '@/styles/themePreset'
 import gettext from './gettext'
-import App from './App.vue'
+import App from '@/App.vue'
 import router from './router'
 import * as Sentry from '@sentry/vue'
